# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API de Órdenes de Compra a Proveedores
  description: >
    Contrato de referencia para la implementación del backend. Este archivo es
    el contrato preciso contra el cual se implementa y se mantiene a mano.
  version: "2026-09-24"
  contact:
    name: Henry Molina
    url: https://henry-molina.dev/
servers:
  - url: http://localhost:8080/api
    description: Entorno local (Docker Compose)
security:
  - bearerAuth: []
...
paths:
...
  /usuarios:
    get:
      tags: [Usuarios]
      summary: Listar usuarios
      operationId: listarUsuarios
      description: >
        Cuenta con acceso al sistema (rol y, si aplica, sucursal asignada).
        Exclusivo de ADMIN: es gestión de identidades y permisos, fuera del
        alcance de cualquier otro rol.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Listado de usuarios
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Usuario" }
    post:
      tags: [Usuarios]
      summary: Crear usuario
      operationId: crearUsuario
      description: >
        Da de alta una nueva cuenta de acceso, asignándole un rol y, si es
        GERENTE_SUCURSAL, la sucursal a la que queda restringida. Exclusivo de
        ADMIN.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UsuarioRequest" }
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usuario" }
  /usuarios/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPathParam"
    put:
      tags: [Usuarios]
      summary: Actualizar usuario
      operationId: actualizarUsuario
      description: >
        Edita una cuenta existente (contraseña, rol, sucursal), o la
        activa/desactiva mediante el campo activo. Exclusivo de ADMIN.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UsuarioRequest" }
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usuario" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT propio emitido por /auth/login (sin OAuth2)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Credencial por proveedor, usada exclusivamente por el endpoint de webhooks
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  schemas:
    Rol:
      type: string
      enum: [ADMIN, COMPRADOR, GERENTE_SUCURSAL]
...
    ListEnvelope:
      type: object
      properties:
        object: { type: string, enum: [list] }
        has_more: { type: boolean }
...
    Usuario:
      type: object
      properties:
        object: { type: string, enum: [usuario] }
        id: { type: integer }
        username: { type: string }
        rol: { $ref: "#/components/schemas/Rol" }
        sucursal_id: { type: integer, nullable: true }
        activo: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UsuarioRequest:
      type: object
      required: [username, rol]
      properties:
        username: { type: string, minLength: 3, maxLength: 60 }
        password:
          type: string
          format: password
          minLength: 8
          description: Obligatorio al crear; en edicion, si se omite se conserva la contraseña actual.
        rol: { $ref: "#/components/schemas/Rol" }
        sucursal_id:
          type: integer
          nullable: true
          description: Obligatorio si rol = GERENTE_SUCURSAL
        activo:
          type: boolean
          description: Omitido en creacion equivale a true; omitido en edicion preserva el valor actual.
...